requirements.txt
langchain
langchain-community
faiss-cpu
pymupdf
streamlit
python-dotenv
sentence-transformers
transformers
torch
